requirements.txt
transformers
torch
Pillow
pytest
